refactor: 텍스트 핸들러 커스텀 훅
